+++ client/dashboard/sites/performance/core-web-vitals/charts/style.scss
