+++ client/dashboard/sites/performance/core-web-vitals/core-web-vitals-accordion.tsx
