+++ client/dashboard/sites/performance/core-web-vitals/metric-tab-bar.tsx
