+++ client/dashboard/sites/performance/core-web-vitals/status-section.scss
