+++ client/dashboard/sites/performance/core-web-vitals/style.scss
